add mailer config to /configs
