+++ app/api/mailer/contact/route.ts
@@ -1,22 +1,5 @@
 import { NextResponse } from 'next/server';
-import nodemailer from 'nodemailer';
-
-const host = process.env.EMAIL_SERVER_HOST;
-const email = process.env.EMAIL_SERVER_USER;
-const pass = process.env.EMAIL_SERVER_PASSWORD;
-const port = process.env.EMAIL_SERVER_PORT;
-
-export const transporter = nodemailer.createTransport({
-  host,
-  port,
-  secure: true,
-  service: 'gmail',
-  auth: {
-    type: 'login',
-    user: email,
-    pass,
-  },
-});
+import { transporter } from '@/configs/mailer';
 
 export async function POST(req: Request) {
   const data = await req.json();
